Implemented cart product notification for the add to cart button
